-Moved theme.ts to its own directory in src.
