+++ src/App.tsx
@@ -4,7 +4,7 @@ import "./App.css";
 import { ApolloProvider } from "@apollo/react-hooks";
 import ApolloClient from "apollo-boost";
 import { ThemeProvider } from "@material-ui/styles";
-import theme from "./theme"
+import theme from "./theme/theme"
 import { ThemeContext } from "./context/contexts";
 
 const client = new ApolloClient({
